src/api/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
anthropic
pyyaml
